change to int instead of bigint

+++ db/migrations/20231017005257_add_origin_weight.sql
@@ -1,6 +1,6 @@
 -- +goose Up
 -- modify "origins" table
-ALTER TABLE "origins" ADD COLUMN "weight" bigint NOT NULL DEFAULT 100;
+ALTER TABLE "origins" ADD COLUMN "weight" int NOT NULL DEFAULT 100;
 
 -- +goose Down
 -- reverse: modify "origins" table
